refactor: Переносит создание store в отдельный файл
